requirement.txt
langchain-google-genai
streamlit
langchain-community
faiss-cpu
pypdf 
python-dotenv
sentence-transformers